Moved kafka-to-pubsub to examples/ directory and updated README.md
