+++ examples/kafka-to-pubsub/build.gradle
@@ -1,0 +1,49 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * License); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an AS IS BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+plugins {
+    id 'java'
+    id 'org.apache.beam.module'
+    id 'com.github.johnrengelman.shadow'
+}
+applyJavaNature(
+        exportJavadoc: false,
+        automaticModuleName: 'org.apache.beam.examples.kafka-to-pubsub')
+provideIntegrationTestingDependencies()
+enableJavaPerformanceTesting()
+
+description = "Apache Beam :: Templates :: Kafka to PubSub"
+ext.summary = """Apache Beam SDK provides a simple, Java-based
+interface for processing virtually any size data. This
+artifact includes Kafka to PubSub template for Google DataFlow"""
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    compile library.java.kafka_clients
+    compile project(path: ":sdks:java:core", configuration: "shadow")
+    compile project(":sdks:java:extensions:google-cloud-platform-core")
+    compile project(":sdks:java:io:google-cloud-platform")
+    compile project(":sdks:java:io:kafka")
+    compile project(":runners:google-cloud-dataflow-java")
+    runtime project(path: ":runners:direct-java", configuration: "shadow")
+    testCompile group: 'junit', name: 'junit', version: '4.12'
+}
+
