+++ examples/kafka-to-pubsub/src/main/resources/kafka_to_pubsub_metadata.json
